requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.2
distlib==0.3.6
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
identify==2.5.19
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pre-commit==3.1.1
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
virtualenv==20.21.0
yarl==1.8.2
